# examples/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stparts Price API",
    "version": "0.1.0"
  },
  "paths": {
    "/jobs": {
      "post": {
        "summary": "Submit",
        "operationId": "submit_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Progress",
        "operationId": "progress_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Progress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/results": {
      "get": {
        "summary": "Results",
        "operationId": "results_jobs__job_id__results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Results"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Accepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "batch_id",
          "status",
          "total"
        ],
        "title": "Accepted"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ItemError"
      },
      "JobRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Batch Id"
          },
          "max_delivery_days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 0.0,
            "title": "Max Delivery Days",
            "default": 2
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/Part"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Parts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "parts"
        ],
        "title": "JobRequest"
      },
      "Part": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Id"
          },
          "oem": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Oem"
          },
          "brand": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Brand"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "oem",
          "brand"
        ],
        "title": "Part"
      },
      "PartResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "found",
              "not_found",
              "error"
            ],
            "title": "Status"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "delivery_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Days"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Id"
          },
          "oem": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Oem"
          },
          "brand": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Brand"
          },
          "currency": {
            "type": "string",
            "const": "RUB",
            "title": "Currency",
            "default": "RUB"
          },
          "parsed_at": {
            "type": "string",
            "title": "Parsed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "id",
          "oem",
          "brand",
          "parsed_at"
        ],
        "title": "PartResult"
      },
      "Progress": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "found": {
            "type": "integer",
            "title": "Found"
          },
          "not_found": {
            "type": "integer",
            "title": "Not Found"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "batch_id",
          "status",
          "total",
          "processed",
          "found",
          "not_found",
          "errors"
        ],
        "title": "Progress"
      },
      "Results": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "max_delivery_days": {
            "type": "integer",
            "title": "Max Delivery Days"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/PartResult"
            },
            "type": "array",
            "title": "Parts"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "batch_id",
          "status",
          "max_delivery_days",
          "expires_at",
          "parts"
        ],
        "title": "Results"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}